+++ bin/image-guard.js
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
-import { styleText } from 'node:util'
 import { runImageGuard } from '../src/index.js'
+import { styleStderr } from '../src/utils.js'
 
 try {
   await runImageGuard()
 } catch (err) {
   if (err.setupFailed) {
-    console.error(styleText('red', err.message))
+    console.error(styleStderr('red', err.message))
   } else {
     console.error('Error running Image Guard:', err)
   }